add GithubAppClientTest#listAccessibleRepositories and refactor other tests
This commit refactors GithubAppClientTest so that it tests the actual
methods in GithubAppClient and adds a test of the new method
`listAccessibleRepositories()`.

Previously the test did not actually test any logic in GithubAppClient
but rather the tests were of the deserialization logic for some of the
types used in the class.

I moved the test related to deserializing an AccessToken to a new
AccessTokenTest, and replaced the other tests so that they use a
MockWebServer to actually be able to test the HTTP requests/responses.