Add path and status code to exception message
Before this change, if you got a RequestNotOkException in a log file, it
wasn't possible to see which request it was that failed.

With this change in place, the path and status code of the exception are
prepended to the exception message, and any logged exceptions will
contain information on what operation it was that failed.