Fix parsing non-protected branches + simplify deserialization
In https://github.com/spotify/github-java-client/pull/82 we
added a custom deserializer for the `Branch` class.

This deserializer didn't handle certain cases properly that the GH API
would return. In particular the case where the `protected` field was set
and no `protection_url` is present would lead to a NPE in the
deserializer.

To avoid this error and make the whole parser more resilient to
different API results, let's reduce the scope of the custom deserializer
to only parse the `protection_url` field instead of the whole `Branch`
class.

There should be no change in runtime behavior apart from fixing the
error cases.